backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
python-dotenv>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
pydantic>=2.10.0
